database - constructor refactor/improvements

+++ Library/BaseInstance.cs
@@ -1,7 +1,7 @@
 ﻿namespace Library
 {
 	// Abstract of the shared properties of Instance or GroupedInstance
-	public abstract class BaseInstance(Entity entity, Space space, string label, LockLevel lockLevel, float spawnWeight)
+	public abstract class BaseInstance(Entity entity, Space space, string label, LockLevel lockLevel, double spawnWeight)
 	{
 		// The Entity which this is an instance of
 		public Entity Entity { get; } = entity;
@@ -17,6 +17,6 @@
 
 		// The expected quantity of the spawn
 		// EG 0.5 for a 50% chance NPC, 2 for junk containing 2 scrap, 1 (Except rarely when Entity.percChanceNone has a value) for a statically placed object
-		public float SpawnWeight { get; } = spawnWeight;
+		public double SpawnWeight { get; } = spawnWeight;
 	}
 }
